cleaned up Card List Features and setup story components properly
